consolidate response processing and eliminate dual-path architecture
Complete integration of response processing logic for cleaner architecture:
- Remove ResponseExtractor class and consolidate functionality into ResponseProcessor
- Move ExtractionResult to response/types.py module for better organization
- Eliminate dual processing paths (with/without response processor)
- Unify all processing through integrated ResponseProcessor approach
- Remove response_processor parameter from processing method signatures
- Simplify ProcessingOptions dataclass removing unused response_processor field

Eliminates code duplication and creates single, consistent processing pathway while maintaining all existing functionality. Provides clean foundation for context caching implementation.
